improve precision

+++ seat-monitor/app/src/main/java/com/novoda/loadgauge/Ads1015.java
@@ -117,7 +117,7 @@ public interface Ads1015 {
     }
 
     interface ComparatorCallback {
-        void onThresholdHit(int valueInMv);
+        void onThresholdHit(float valueInMv);
     }
 
     class Factory {
